+++ test/repositories/local-file-test.js
@@ -9,19 +9,13 @@ var assert = require('assert'),
     fs = require('fs'),
     path = require('path'),
     exec = require('child_process').exec,
-    eyes = require('eyes'),
     vows = require('vows'),
     helpers = require('../helpers'),
     haibu = require('../../lib/haibu');
 
-var ipAddress = '127.0.0.1', 
-    port = 9000, 
-    app = {
+var app = {
        "name": "test",
        "user": "marak",
-       "directories": {
-         "home": "hellonode"
-       },
        "repository": {
          "type": "local",
          "directory": path.join(__dirname, '..', 'fixtures', 'repositories', 'local-file'),
@@ -40,15 +34,13 @@ vows.describe('haibu/repositories/local-file').addBatch(helpers.requireInit()).a
       "should be a valid repository": function (localFile) {
         assert.instanceOf(localFile, haibu.repository.Repository);
         assert.isFunction(localFile.init);
-        assert.isFunction(localFile.exists);
-        assert.isFunction(localFile.update);
         assert.isFunction(localFile.fetch);
       },
       "the fetch() method": {
         topic: function (localFile) {
           localFile.fetch(this.callback);
         },
-        "use the local filesystem": function (err, localFile) {
+        "should find the local app directory": function (err, localFile) {
           try {
             assert.isNotNull(fs.statSync(localFile));
           }
@@ -57,21 +49,11 @@ vows.describe('haibu/repositories/local-file').addBatch(helpers.requireInit()).a
             assert.isNull(ex);
           }
         }
-      }
-    }
-  }
-}).addBatch({
-  "When using haibu": {
-    "an instance of the LocalFile repository": {
-      topic: function () {
-        return haibu.repository.create(app);
-      },
-      "should be a valid repository": function (localFile) {
-        assert.instanceOf(localFile, haibu.repository.Repository);
       },
       "the init() method": {
         topic: function (localFile) {
           var self = this;
+          if (!(localFile instanceof haibu.repository.Repository)) return localFile;
           exec('rm -rf ' + path.join(localFile.appDir, '*'), function(err) {
             localFile.mkdir(function (err, created) {
               if (err) self.callback(err);
